2025 04 08 17:38 주문/카트 엔티티 디티오 레포지토리등...

+++ src/main/java/com/sixthsense/hexastay/dto/StoremenuDTO.java
@@ -28,7 +28,7 @@ public class StoremenuDTO {
     private String storemenuName;
 
     //룸서비스외부 업체상품 가격
-    private Long storemenuPrice;
+    private Integer storemenuPrice;
 
     //룸서비스외부 업체상품 설명
     private String storemenuContent;
